requirements.txt
numpy
matplotlib
opencv-contrib-python
Pillow
scikit-learn
docker
kubernetes
ultralytics
flask